requirements.txt
python-telegram-bot==20.7
openai
fastapi
uvicorn
sqlite3